+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.3.1 under development
 
+- Enh #226: Check empty string as key in command map in `CommandLoader` validation (@vjik)
 - Bug #224: Fix typo in `UNAVAILABLE` exit code reason (@vjik)
 
 ## 2.3.0 January 23, 2025